Adding API calls to authenticate user and display fundraisers
